feat: add config helpers for managing Codex and MCP server configurations
