+++ UnityMcpBridge/Editor/Helpers/McpConfigFileHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f69ad468942b74c0ea24e3e8e5f21a4b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
